+++ apps/storybook/.storybook/main.ts
@@ -2,6 +2,7 @@ import { dirname, join, resolve } from 'node:path';
 import { env } from 'node:process';
 
 import type { StorybookConfig } from '@storybook/react-vite';
+import type { PropItem } from 'react-docgen-typescript';
 
 const config: StorybookConfig = {
   typescript: {
@@ -20,6 +21,14 @@ const config: StorybookConfig = {
       include: [resolve(__dirname, '../../../packages/react/**/**.tsx')], // <- This is the important line.
       shouldExtractLiteralValuesFromEnum: true,
       shouldRemoveUndefinedFromOptional: true,
+      propFilter: (prop: PropItem) => {
+        // Keep the default filter logic from storybook:
+        // https://github.com/storybookjs/storybook/blob/40523765403480c4065373664553bb6cbdf9543d/code/core/src/core-server/presets/common-preset.ts#L169C5-L169C98
+        const defaultLogicFromStorybook = prop.parent
+          ? !/node_modules/.test(prop.parent.fileName)
+          : true;
+        return defaultLogicFromStorybook && prop.name !== 'popovertarget'; // Skip popovertarget @types/react-dom patch}
+      },
     },
   },
   stories: [