+++ packages/react/src/components/HelpText/HelpText.tsx
@@ -1,11 +1,9 @@
 import type { Placement } from '@floating-ui/utils';
 import cl from 'clsx/lite';
+import { forwardRef, useId, useState } from 'react';
 import type { ButtonHTMLAttributes } from 'react';
-import { useState } from 'react';
 
-import type { PortalProps } from '../../types/Portal';
-import { Popover } from '../Popover';
-import type { PopoverRootProps } from '../Popover/PopoverRoot';
+import { Popover, type PopoverProps } from '../Popover';
 import { Paragraph } from '../Typography/Paragraph';
 
 import { HelpTextIcon } from './HelpTextIcon';
@@ -19,72 +17,74 @@ export type HelpTextProps = {
    * Size of the helptext
    * @default md
    */
-  size?: PopoverRootProps['size'];
+  size?: PopoverProps['size'];
   /**
    * Placement of the Popover.
    * @default 'right'
    */
   placement?: Placement;
-} & PortalProps &
-  ButtonHTMLAttributes<HTMLButtonElement>;
+} & ButtonHTMLAttributes<HTMLButtonElement>;
 
-export const HelpText = ({
-  title,
-  placement = 'right',
-  portal,
-  size = 'md',
-  className,
-  children,
-  ...rest
-}: HelpTextProps) => {
-  const [open, setOpen] = useState(false);
+export const HelpText = forwardRef<HTMLButtonElement, HelpTextProps>(
+  function HelpText(
+    {
+      title,
+      placement = 'right',
+      size = 'md',
+      className,
+      children,
+      ...rest
+    }: HelpTextProps,
+    ref,
+  ) {
+    const [open, setOpen] = useState(false);
+    const randomId = useId();
 
-  return (
-    <>
-      <Popover.Root
-        variant='info'
-        placement={placement}
-        size={size}
-        portal={portal}
-        open={open}
-        onClose={() => setOpen(false)}
-      >
-        <Popover.Trigger asChild variant='tertiary'>
-          <button
+    return (
+      <>
+        <button
+          className={cl(
+            `ds-helptext--${size}`,
+            'ds-helptext__button',
+            `ds-focus`,
+            className,
+          )}
+          onClick={() => setOpen(!open)}
+          // @ts-ignore
+          popovertarget={randomId}
+          ref={ref}
+          {...rest}
+        >
+          <HelpTextIcon
+            filled
             className={cl(
-              `ds-helptext--${size}`,
-              'ds-helptext__button',
-              `ds-focus`,
+              `ds-helptext__icon`,
+              `ds-helptext__icon--filled`,
               className,
             )}
-            aria-expanded={open}
-            onClick={() => setOpen(!open)}
-            {...rest}
-          >
-            <HelpTextIcon
-              filled
-              className={cl(
-                `ds-helptext__icon`,
-                `ds-helptext__icon--filled`,
-                className,
-              )}
-              openState={open}
-            />
-            <HelpTextIcon
-              className={cl(`ds-helptext__icon`, className)}
-              openState={open}
-            />
-            <span className={`ds-sr-only`}>{title}</span>
-          </button>
-        </Popover.Trigger>
+            openState={open}
+          />
+          <HelpTextIcon
+            className={cl(`ds-helptext__icon`, className)}
+            openState={open}
+          />
+          <span className={`ds-sr-only`}>{title}</span>
+        </button>
+        {/* TODO: Why is popover wrapped in paragraph here? */}
         <Paragraph size='md' asChild>
-          <Popover.Content className='ds-helptext__content'>
+          <Popover
+            id={randomId}
+            className='ds-helptext__content'
+            onClose={() => setOpen(false)}
+            open={open}
+            placement={placement}
+            size={size}
+            variant='info'
+          >
             {children}
-          </Popover.Content>
+          </Popover>
         </Paragraph>
-      </Popover.Root>
-    </>
-  );
-};
-
-HelpText.displayName = 'HelpText';
+      </>
+    );
+  },
+);