+++ packages/react/CHANGELOG.md
@@ -1,5 +1,52 @@
 # Change Log
 
+## 1.0.0-next.33
+
+### Patch Changes
+
+- Pagination: Use data attrs instead of class names ([#2395](https://github.com/digdir/designsystemet/pull/2395))
+
+- Badge: Style using css attributes ([#2391](https://github.com/digdir/designsystemet/pull/2391))
+
+- TableHeaderCell: Remove `sortable` prop, `sort` now handles this ([#2393](https://github.com/digdir/designsystemet/pull/2393))
+
+- dropdownmenu: Style using data attributes ([#2387](https://github.com/digdir/designsystemet/pull/2387))
+
+- List: Remove `List.Root` and `List.Heading`, which changes API ([#2348](https://github.com/digdir/designsystemet/pull/2348))
+
+- Alert, Avatar, Button, Divider, Link: Use data-attributes for variant, size and color and move icons to CSS ([#2313](https://github.com/digdir/designsystemet/pull/2313))
+
+- Box: Remove component ([#2372](https://github.com/digdir/designsystemet/pull/2372))
+
+- Popover: ([#2369](https://github.com/digdir/designsystemet/pull/2369))
+
+  - Rename `<Popover.Root>` to `<Popover.Context>`
+  - use Popover API, allowing `<Popover>` to be used without `Popover.Context`
+  - Remove `portal` prop
+
+- Select: Rename from `NativeSelect` ([#2404](https://github.com/digdir/designsystemet/pull/2404))
+
+- Table: Set sort button type to prevent form submit ([#2402](https://github.com/digdir/designsystemet/pull/2402))
+
+- Heading: default level is now 2 ([#2378](https://github.com/digdir/designsystemet/pull/2378))
+
+- Select: ([#2415](https://github.com/digdir/designsystemet/pull/2415))
+
+  - Add Select.Option and Select.Optgroup compond components
+  - Remove `multiple` prop
+
+- Accordion: Now uses details and summary HTML elements ([`5d1c5062b526e6829c322ce66c6df08568bb9f63`](https://github.com/digdir/designsystemet/commit/5d1c5062b526e6829c322ce66c6df08568bb9f63))
+
+- Spinner: Style using data attributes ([#2390](https://github.com/digdir/designsystemet/pull/2390))
+
+- Avatar: new component ([#2312](https://github.com/digdir/designsystemet/pull/2312))
+
+- Tag: Make neutral default color in CSS ([#2397](https://github.com/digdir/designsystemet/pull/2397))
+
+- Card: Use data attrs ([#2398](https://github.com/digdir/designsystemet/pull/2398))
+
+- Combobox: fix virtual combobox having large gap between items ([#2376](https://github.com/digdir/designsystemet/pull/2376))
+
 ## 1.0.0-next.32
 
 ## 1.0.0-next.17